fix: Move policy param setter items to footer
