+++ app/src/components/policyParameterSelectorFrame/Header.tsx
@@ -1,36 +1,9 @@
-import { useDispatch } from 'react-redux';
-import { Button, Group, Text } from '@mantine/core';
-import { clearPolicy } from '@/reducers/policyReducer';
-import { FlowComponentProps } from '@/types/flow';
+import { Text } from '@mantine/core';
 
-export default function PolicyParameterSelectorHeader({
-  onNavigate,
-  onReturn,
-}: FlowComponentProps) {
-  // TODO: Determine how to handle policy number
-
-  const dispatch = useDispatch();
-
-  function handleNext() {
-    // Dispatch an action to move to the next step
-    onNavigate('next');
-  }
-
-  function handleCancel() {
-    // Clear policy state and return to the previous step
-    dispatch(clearPolicy());
-    onReturn();
-  }
-
+export default function PolicyParameterSelectorHeader() {
   return (
-    <Group justify="space-between" align="center">
-      <Button variant="default" onClick={handleCancel}>
-        Cancel
-      </Button>
-      <Text fw={700}>Policy #NUMBER</Text>
-      <Button variant="filled" onClick={handleNext}>
-        Next
-      </Button>
-    </Group>
+    <Text fw={700} size="lg">
+      Policy Parameters
+    </Text>
   );
 }
